avoid parsing /dev/cua01 as default value
This is a minor tweak with an important side effect for the parser [1]
that read LCDd.conf to generate LCDproc model [2]

Example like '#Device=/dev/cua01' is interpreted as a default value. In
this case, this is not consistent with the default value indicated in
the line above (/dev/ttyS1)

With the extra space, the parser ignore the value.

[1] https://github.com/dod38fr/config-model-lcdproc/blob/master/script/lcdconf2model.pl
[2] https://ddumont.wordpress.com/2011/07/03/generate-a-configuration-editor-from-a-config-template-file-with-perl-lcdproc-example/