h44780: Remove duplicate code
The code had been moved to timing.h long ago, but got reintroduced by
some merge accident in 2001 it seems.

The code did nothing by default, was overridden by the inclusion of
timing.h later in the file and would be interacting strangely with the
code therein.

This partially addresses issue #56.

Beported-by: Philip Prindeville <philipp_subx@redfish-solutions.com>
Signed-off-by: Harald Geyer <harald@ccbib.org>