sdeclcd: don't rely on STRB bit being 1/0 when manipulating backlighting
This code works on the lucky coincidence that backlighting is controlled
by the STRB bit whose value is 1, and that this function is called with
a 0 or 1 value to reset/set backlighting.

If I called sdec_write(..., 99) that would also be a valid way of
turning on the backlight, since that argument is ultimately zero or
non-zero in how it's interpreted.  But that would then mask in the
wrong value into the control register.

So, test backlight for being non-zero, but then assign the right bit
(which happens to be 1 or 0) into the variable we then combine to
generate the correct register value.

If we were running on different hardware were the 2 bit controlled
the backlighting, obvious this wouldn't work... so it's a lucky
coincidence that it does.

Signed-off-by: Philip Prindeville <philipp@redfish-solutions.com>