Fix errors in buffer sizes for ESC/POS & cd5220. Update for consistency.
The two bytes used to specify the range of character codes to replace
were not taken into account when calculating the size of the command
buffer required for these protocols, and could lead to buffer overflows
for the command buffer.

Changed test for custom character support by display to be more
consistent across the two protocol drivers.

Also corrected a minor documentation mistake, the amount of glyphs we
need to send is equivalent to the (endcode - startcode) + 1 as the
character replacement range is inclusive. No functionality changes were
required for this change.