clients/lcdproc/cpu: Make CPU screen deal better with screens with an odd width
This commit modifies the lcdproc CPU screen to put the column separating
space in the middle in the case of a LCD panel with an odd width in
characters. Before this commit the column separating the space would be
1 character on the left of the actual display center.

Tested on a mx5000 LCD panel which has a width of 17 characters.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>