clients/lcdproc/cpu: Make CPU screen deal better with small screens
The 4+ lines CPU screen so far has been assuming that the width is at
least 20 chars. This is not always the case.

This commit trims characters of the labels when printing them, making sure
there always is whitespace between the label and the actual percentage.

This also fixes how much of the label is visible being different for
percentages <= 9.9% versus larger percentages.

This makes the lines look like this on a 17 char wide panel:

 * +-----------------+
 * |## CPU 51.9%:  #@|
 * |Us 46.0% Ni  0.0%|
 * |Sy  5.9% Id 48.1%|
 * |0%------     100%|
 * +-----------------+

And like this on a 16 char wide panel:

 * +----------------+
 * |## CPU 51.9%: #@|
 * |U 46.0% Ni  0.0%|
 * |S  5.9% Id 48.1%|
 * |0%------    100%|
 * +----------------+

Tested on a mx5000 LCD panel which has a width of 17 characters.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>