clients/lcdproc/mem: Speedup mem screen title setting code
Instead of setting a title and then overwriting part of the title with
a string overwriting the title location, simply printf the actual combined
string we want the title to become and send that as the title.

This results in less data being send to the LCD and with slow panels the
hash symbols padding out the title may be briefly seen before the "subtitle"
overwrites them which is not pretty.

Also on some (rare) panels overwriting text without clearing the screen in
between leads to the hashes and the SWAP text overstriking one each-other
which looks really ugly.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>