clients/lcdproc/mem: Make mem screen deal better with small screens
The 4+ lines mem screen already has some code to deal with screens which
are less then 20 chars wide. But the code printing the center "Totl" and
"Used" labels so far has been assuming that the width is at least 20 chars.

This commit trims characters of the labels when printing them, making sure
there always is whitespace between the label and the actual memory values
being printed.

While at it also rename the string widget for the Free label from "used"
to "free" since having a label named "used" containing a string "Free"
is confusiong.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>