clients/lcdproc/util: Fix sprintf_memory sometime sprinting 7 chars
sprintf_memory is supposed to generate max 6 chars output, but on my
workstation with a swapsize pof 9.999996185 it was chosing to use
"%.3f%s" as format string and then after rounding to the requested
precision sprintf ends up formatting this as "10.000G" consuming 7
chars instead of 6.

This commit fixes this by comparing to the maximum value which we can
print after rounding to the precision we plan to use.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>