server: Add a new pbar (percentage-bar) widget
Various lcdproc clients / screens draw a horizontal bar between a begin and end
label to indicate how much percent of a resource is used. Examples of this
are CPU load, memory use and swap usage bars.

More specifically the CPU load bar in the lcdproc CPU screen in 4 line mode
is bracketed between "0%" and "100%" texts and the mem and swap usage bars
in the lcdproc memory screen in 4 line mode are bracketed between "E" and "F".
In 2 line screens where there is little space, these horizontal-bars are
typically bracketed between "[" and "]" to take up as little space as possible
while still having an indicator where the bar ends / where 100% usage is.

Having a rectangular outline / background which is filled horizontally from
left to right (e.g. your typical progressbar) is a standard UI element in
lots of software. So some LCD screens / drivers are capable of drawing
such percentage-bars natively, including a rectangular outline / background
indicating where the begin and end of the bar are.

This commit introduces a new pbar (percentage-bar) widget for this often seen
pattern. Having a separate widget for this has a couple of advantages:

1) If there are no specific begin and end labels (just the '[' and ']') the
begin and end-labels can be left out completely since the begin and end is
already marked by the pbar background, leaving more width for the actual pbar.

2) Some drivers cannot draw hbars but they can draw pbars, the vbar fallback
implementation which just draws a couple of '-' chars is not really pretty.
Using a native pbar looks much better in these cases.

3) By letting the driver render the labels, the driver can optionally make
them take up less space, by e.g. using a smaller font, or just leaving them
out all together if space is really tight.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>