server/drivers/mx5000: Rename hbar implementation to pbar implementation
libmx5000's mx5000_sc_add_progress_bar() function draws a progressbar,
not a horizontalbar. This means a background gets drawn for the non-filled
part of the bar, which is not what a hbar method is supposed to do. The
driver's hbar method should only render / draw the filled part of the hbar.

This commit renames the hbar implementation into a pbar implementation,
as mx5000_sc_add_progress_bar() does actually draw a pbar.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>