clients/lcdproc: Rework 2 line mem display gauge width and offset code
The 2 line mem display has 3 different cases for handling the mem/swap
gauges depending on the lcd-width:

1) lcd_wid >= 20, draw gauges + separating spaces before and after the gauge
2) lcd_wid >= 18, draw gauges, without any spaces before and after the gauge
3) lcd_wid < 18, no gauges

This commits makes 2 changes to this:

i. It reworks the very hard to read old code for calculating the gauge_wid
and offs for the 3 cases into 3 clearly separate cases, so that it is much
easier to see what is going on.
ii. The old code assumes that lcd_wid is always even, but with some special
LCDs it might be odd. Specifically this commit changes the limit for not
drawing any gauges from 18 to 17, where we have just enough room for a gauge,
resulting in the following 3 new cases:

1) lcd_wid >= 20, draw gauges + separating spaces before and after the gauge
2) lcd_wid >= 17, draw gauges, without any spaces before and after the gauge
3) lcd_wid < 17, no gauges

Signed-off-by: Hans de Goede <hdegoede@redhat.com>