server/drivers/g15: Fix chr function overwriting text above/below the char
The g15 driver uses 5 lines of 8 pixels height. The G15_TEXT_LARGE font
we use for this has a height of up to 8 pixels per character. Uppercase
characters and numbers use the top 7 pixels, and lowercase chars mostly
use the middle 6 pixels.

But there are some exceptions, the 'g' and 'y' lower-case chars use the
8th row of pixels for the bottom parts / curve (descender) of the glyph.

Because some chars do use the full 8 pixels height; and to avoid these
pixels touching other rendered content, the g15r_renderCharacterLarge()
clears a background area for the char of 10 pixels height; so it also
clears a line above and below the char, potentially removing parts of the
line above / below the line being rendered.

When first rendering line 1, then line 2, etc. this results in the 8th
pixel-row of the previous line being cleared; this is mostly empty, but
this removes the descender from the lower-case 'g' and 'y' glyphs.
If OTOH first line 3 is rendered and then line 2, as the 4 line mem screen
of the lcdproc client does, then the first pixel row of line 3 gets cleared
which looks quite bad.

This commit fixes this by instead of calling g15r_renderCharacterLarge
loading the font ourselves and then using g15r_renderG15Glyph and telling
g15r_renderG15Glyph to not clear the background.

Since we may be requested to overwrite something written at the location
before, we now need to manually clear the blackground before rendering the
glyph. This commit adds the necessary clearing of the background, but
only for a 8x8 pixel area, so that the lines above/below the current line
do not get touched.

This fixes sometimes the first pixel row and often the 8th pixel-row of
adjacent lines getting erased.

This does mean that a 'g' or 'y' above an upper-case letter; a number;
or a 't' will touch the glyph below it. Still this is much better then
misrendering things. And a follow-up patch will mostly fix this too.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>