server/drivers/g15: Fix of by 1 error in vbar method
g15r_pixelBox draws a box from pixel y1 through y2, so if we want to
draw a 8 pixel long vbar starting at the top of the screen we would pass
y1 = 0 and y2 = 7, not y1 = 0 and y2 = 8, as that will draw a 9 pixel long
vbar.

This commit fixes this by adding the missing - 1 to the y2 calculation.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>