server/drivers/g15: Add an empty pixel row between the text lines.
The descender parts of a lower-case 'g' or 'y' glyph will touch the glyph
below it, if the glyph below is an upper-case letter; a number; or has an
ascender like the 't' glyph.

We have 3 unused pixel rows at the end of the LCD, this commit uses those
to fix this for the first 4 lines of the display, by adding an empty row
betweent the lines.

Unfortunately there is no space to also add an empty pixel row between the
4th and 5th lines, but many screens only use 4 lines, so this does fix the
descender touching the row below issue in many cases.

This also makes us better use the whole height of the LCD.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>