server/drivers/g15: Remove bogus init of buffer[0] with G15_LCD_WRITE_CMD
G15_LCD_WRITE_CMD needs to send to the g15daemon *once* not as part of
every frame, since we pass G15_G15RBUF to new_g15_screen, this is already
done for us by new_g15_screen.

buffer[0] contains the first 8 pixels of the first column of the display
and setting this to G15_LCD_WRITE_CMDn makes no sense what so ever.
This code was not causing problems because we do a g15r_clearScreen later
on which undoes it, but it is totally bogus, so lets remove it.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>