server/drivers: Add libhidraw.a helper library
Add a helper library for accessing /dev/hidraw# devices. The main purpose
of this helper-lib is to allow drivers to specify a list of supported
products and then have the lib figure out which /dev/hidraw# to open.

This is necessary because the numbering of the /dev/hidraw# devices
depends on the enumeration order of the USB devices and this order may
differ from one boot to the next.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>