server/drivers/g15: Add support for directly accessing the LCD
Add support to the g15 driver for it directly accessing the LCD through the
/dev/hidraw interface.

There are several reasons to do this:
1. LCDd itself is a daemon for driving LCD screens, it really should not
depend on yet another daemon for something which it can easily do itself.

2. g15daemon detaches the kernel driver from the USB interface it used,
this causes several problems:

Problem 1. On the G510 the same USB interface is used for both the media
keys (play/pause prev/next song, stop, volume up/down/mute) as well as for
the LCD panel. This means that g15daemon causes the media keys on the G510
model to stop working.

Problem 2. The extra macro keys on these keyboards are really just extra keys,
all the macro stuff is done in software on the host. There are kernel patches
pending upstream which move the handling of these extra keys to the kernel,
where it belongs. These patches make the kernel to send KEY_MACRO# events for
these keys. This is part of an upstream effort to create a single, generic,
vendor agnostic solution for keyboards with macro-keys, see:
https://github.com/libratbag/libratbag/issues/172  g15daemon conflicts with
this, esp. the unbinding of the kernel driver it does.

Therefor this commit adds support to the g15 driver to optionally talk
directly to the keyboard without needing g15daemon. This defaults to off
to preserve current behavior for existing users.

Since we now no longer need g15daemon, this commit also allows building
the g15 driver without libg15daemon_client in which case it will always
use hidraw access.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>