server/drivers/g15: Fix heap corruption
The g15canvas struct defined by libg15render.h has a number of optional
struct members at the end which are only declared when TTF_SUPPORT is
defined.

When libg15render is build with TTF_SUPPORT it will unconditionally
write various things to these members, causing heap corruption when
TTF_SUPPORT is not defined when including libg15render.h.

This really is a bug in libg15render, it should really generate the
header during buildtime so that the installed header always matches
the config used during building.

But fixing that is outside of our control. As a workaround we can
always define TTF_SUPPORT on our side. We allocate the canvas, so if
it is bigger the necessary because libg15render was no build with
TTF_SUPPORT, then  we are just allocating some extra memory without any
bad side effects.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>