server/input: Fix NULL pointer deref
Commit 2017fdaaf0fa ("server/input: Cleanup") breaks the case where
input_find_key() returns a KeyReservation with a NULL client.

This happens when using lcdproc compiled with --enable-lcdproc-menus
as single client and then pressing the MenuKey, causing LCDd to crash
as soon as the MenuKey is pressed.

This commit re-adds the check for kr->client being NULL, fixing the crash.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>