server/menu: Fix wrong menu scroll when wrapping on pressing Up
When pressing Up when we are already at the top menu entry, we wrap to
the bottom entry.

Before this commit we used to unconditionally set menu.scroll to
"menu->data.menu.selector_pos + 2 - display_props->height" when wrapping
to the bottom entry. This is wrong if the menu has less entries then the
screen is high, causing menu.scroll empty lines to show above the top
"LCDproc Menu" title widget after the wrap.

This commit fixes this by setting menu.scroll to 0 when the menu has
less entries then the displays height.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>