server/drivers/mx5000: Drop custom input handling
There are kernel patches pending upstream which add proper evdev codes for
the enu buttons below the status LCD-s builtin to some keyboards.

Combined with recent improvements to LCDd's linux_input driver this means
we can use LCDd's linux_input driver instead of using custom input handling
in the mx5000 driver. Also the hardcoded scancodes used by the custom
handling are only correct for the mx5000 and not for the also supported
mx5500 model.

So lets remove this custom handling before we do a release which includes
this custom handling and people may start to rely on it.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>