server/drivers/linux_input: Only use defaults when there is no user config
Before this commit the linux_input event-code to event-string mapping
was hardcoded for ESC, Enter and the cursor-keys and there was no way
to override the defaults, or ignore these keys. Since the defaults where
checked before checking for user configured mappings.

This commit fixes this by first checking if there are user configured
mappings and if those are present then only using the user configured
mappings.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>