server/drivers/linux_input: Consume all events until we find a valid keypress
Since we are only polling at framerate times per second, which with some
slow LCDs might be quite slow, we may build a backlog of events.

This is especially a problem when the input device we use also generates
a lot of events which are not for us, such as happens e.g. with the mx5000
and mx5500 keyboards where there is only 1 input device for both the
regular keys and the lcd-menu buttons.

This can cause a huge amount of lag, this commit fixes this by consuming
all events until we find a valid keypress; or there are no more events.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>