clients/lcdproc: MemTop: be smarter about how much processes to show
Instead of hardcoding the number of processes to 5, which only works
well for LCDs with 4 lines or less, do the following:

On LCDs with 4 lines or less show 5 processes (no change)

On LCDs with 5 lines we have 4 free lines, showing 5 processes with
scrolling here makes things look like we scroll 1 line down and then 1
line back up again, which looks wrong (*). So show 4 processes here.

On LCDs with 6 lines or more, show as many processes as will fit so that
we use the entire screen.

Note that the last 2 cases both come down to showing as many lines as
will fit.

*) This happens e.g. on the LCD panel in various Logitech keyboards, which
fits 5 lines.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>