server/drivers/hd44780: Fill spi_ioc_transfer structure completely
Current kernels have more fields in spi_ioc_transfer than the ones that are
explicitly set by the mcp23s17_read_reg() and mcp23s17_write_reg() functions.
When some of the structure fields are not initialized, calls to ioctl() may
return EINVAL and SPI transactions will not be executed.

Clearing the structure before filling any field is an effective way of
initializing the structure, while ensuring compatibility with a wide range of
kernel versions.