MtxOrb: add support for full range of baud rates
These can be changed from the default and will be remembered by the
display so we only need to support the setting of the baud rates on the
serial interface. It is up to the user to configure this correctly as
there is no way to know the correct baud rate otherwise.