hd44780: Improve full screen refresh
The 4-bit connection types are vulnerable to loosing transmission
state (which nibble the display expects next) in EMI-rich environments.
Thus, when doing a regular full refresh (by default disabled), we now
check for/call a function that does a full reset of the display.

This commit also implements such a function for the GPIO-based
connection types, where this problem is most likely to happen.

Signed-off-by: Harald Geyer <harald@ccbib.org>